% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFinal.R
\name{executeIncidencePrevalenceFinal}
\alias{executeIncidencePrevalenceFinal}
\title{Execute the CHAPTER study (ATLAS-compatible version)}
\usage{
executeIncidencePrevalenceFinal(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  writePrefix = NULL,
  outputFolder,
  databaseId,
  readCohorts = TRUE,
  createAllergyCohorts = TRUE,
  createCancerCohorts = TRUE,
  createCardioCohorts = TRUE,
  createGeneralCohorts = TRUE,
  createPulmoCohorts = TRUE,
  createDenominator = TRUE,
  calculateIncidence = TRUE,
  continueOnIncidenceError = TRUE,
  minCellCount = 5
)
}
\arguments{
\item{connectionDetails}{An object as created by
\code{\link[DatabaseConnector]{createConnectionDetails}}.}

\item{cdmDatabaseSchema}{Schema name where OMOP CDM resides (e.g. 'cdm.dbo').}

\item{cohortDatabaseSchema}{Schema name where intermediate cohorts will be stored.}

\item{writePrefix}{Optional prefix for tables to be created in
\code{cohortDatabaseSchema}.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{readCohorts}{A boolean to choose whether to read and instantiate the initial cohorts
from the available jsons.}

\item{createAllergyCohorts}{A boolean to choose whether to create allergy cohorts.}

\item{createCancerCohorts}{A boolean to choose whether to create cancer cohorts.}

\item{createCardioCohorts}{A boolean to choose whether to create cardiovascular cohorts.}

\item{createGeneralCohorts}{A boolean to choose whether to create healthcare utilization and mortality cohorts.}

\item{createPulmoCohorts}{A boolean to choose whether to create pulmonary cohorts.}

\item{createDenominator}{A boolean to choose whether to create the denominator cohort table.}

\item{calculateIncidence}{A boolean to choose whether to calculate incidence.}

\item{continueOnIncidenceError}{A boolean to choose whether to continue if one outcome incidence query fails.}

\item{minCellCount}{Minimum allowed count to report; smaller cells are suppressed.}
}
\description{
Execute the CHAPTER study (ATLAS-compatible version)
}
\details{
This function executes the incidence analyses of the CHAPTER study using only
the standard ATLAS stack: DatabaseConnector, SqlRender, CohortGenerator, etc.
}
