% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIncidenceResults.R
\name{getIncidenceResults}
\alias{getIncidenceResults}
\title{Get the incidence results of the CHAPTER study (ATLAS-compatible)}
\usage{
getIncidenceResults(
  connection,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTables,
  analyses,
  latestDataAvailability,
  resultsDir,
  createDenominator = TRUE,
  calculateIncidence = TRUE,
  continueOnIncidenceError = TRUE,
  minCellCount
)
}
\arguments{
\item{connection}{A connection object from DatabaseConnector::connect().}

\item{cdmDatabaseSchema}{CDM schema name.}

\item{cohortDatabaseSchema}{Cohort schema name.}

\item{cohortTables}{Character vector of cohort table names to summarize.}

\item{analyses}{A tibble specifying the incidence analyses to
conduct, retrieved from the settings folder}

\item{latestDataAvailability}{Date of the latest data availability.}

\item{resultsDir}{Directory where result CSVs are written.}

\item{createDenominator}{A boolean to choose whether to create the denominator cohort table.}

\item{calculateIncidence}{A boolean to choose whether to calculate incidence.}

\item{continueOnIncidenceError}{A boolean to choose whether to continue if one outcome incidence query fails.}

\item{minCellCount}{Minimum allowed count to report.}
}
\description{
Get the incidence results of the CHAPTER study (ATLAS-compatible)
}
\details{
This function performs incidence calculations and basic cohort
characterisation using SQL and DatabaseConnector, without CDMConnector
or IncidencePrevalence.
}
